Code structure adding plugins indepent CMakeLists and config file
