+++ as2_state_estimator/include/as2_state_estimator/as2_state_estimator.hpp
@@ -1,0 +1,97 @@
+// Copyright 2024 Universidad Politécnica de Madrid
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//    * Redistributions of source code must retain the above copyright
+//      notice, this list of conditions and the following disclaimer.
+//
+//    * Redistributions in binary form must reproduce the above copyright
+//      notice, this list of conditions and the following disclaimer in the
+//      documentation and/or other materials provided with the distribution.
+//
+//    * Neither the name of the Universidad Politécnica de Madrid nor the names of its
+//      contributors may be used to endorse or promote products derived from
+//      this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+
+/**
+* @file as2_state_estimator.hpp
+*
+* An state estimation server for AeroStack2
+*
+* @authors David Pérez Saura
+*          Rafael Pérez Seguí
+*          Javier Melero Deza
+*          Miguel Fernández Cortizas
+*          Pedro Arias Pérez
+*/
+
+#ifndef AS2_STATE_ESTIMATOR__AS2_STATE_ESTIMATOR_HPP_
+#define AS2_STATE_ESTIMATOR__AS2_STATE_ESTIMATOR_HPP_
+
+#include <tf2_ros/buffer.h>
+#include <tf2_ros/static_transform_broadcaster.h>
+#include <tf2_ros/transform_broadcaster.h>
+#include <tf2_ros/transform_listener.h>
+#include <filesystem>
+#include <memory>
+#include <pluginlib/class_loader.hpp>
+#include <rclcpp/rclcpp.hpp>
+#include <geometry_msgs/msg/pose_stamped.hpp>
+#include <geometry_msgs/msg/twist_stamped.hpp>
+#include <nav_msgs/msg/odometry.hpp>
+
+#include <as2_core/names/topics.hpp>
+#include <as2_core/node.hpp>
+#include <as2_core/utils/frame_utils.hpp>
+#include <as2_core/utils/tf_utils.hpp>
+
+#include "plugin_base.hpp"
+
+namespace as2_state_estimator
+{
+
+class StateEstimator : public as2::Node
+{
+public:
+  explicit StateEstimator(const rclcpp::NodeOptions & options = rclcpp::NodeOptions());
+  ~StateEstimator() {}
+
+private:
+  std::filesystem::path plugin_name_;
+  std::shared_ptr<pluginlib::ClassLoader<as2_state_estimator_plugin_base::StateEstimatorBase>>
+  loader_;
+  std::shared_ptr<as2_state_estimator_plugin_base::StateEstimatorBase> plugin_ptr_;
+  std::shared_ptr<tf2_ros::TransformBroadcaster> tf_broadcaster_;
+  std::shared_ptr<tf2_ros::StaticTransformBroadcaster> tfstatic_broadcaster_;
+  std::shared_ptr<tf2_ros::Buffer> tf_buffer_;
+  std::shared_ptr<tf2_ros::TransformListener> tf_listener_{nullptr};
+
+private:
+  /**
+   * @brief Modify the node options to allow undeclared parameters
+   */
+  static rclcpp::NodeOptions get_modified_options(const rclcpp::NodeOptions & options)
+  {
+    // Create a copy of the options and modify it
+    rclcpp::NodeOptions modified_options = options;
+    modified_options.allow_undeclared_parameters(true);
+    modified_options.automatically_declare_parameters_from_overrides(true);
+    return modified_options;
+  }
+};  // class StateEstimator
+}  // namespace as2_state_estimator
+
+#endif  // AS2_STATE_ESTIMATOR__AS2_STATE_ESTIMATOR_HPP_
