--- as2_state_estimator/plugins/ground_truth/config/config_default.yaml
@@ -1,5 +1,0 @@
-/**:
-  ros__parameters:
-    use_ignition_tf: False
-    use_gps: false
-    set_origin_on_start: false
