--- as2_state_estimator/plugins/raw_odometry/config/config_default.yaml
@@ -1,11 +1,0 @@
-/**:
-  ros__parameters:
-    odometry_topic: "sensor_measurements/odom"
-    use_gps: false
-    set_origin_on_start: false
-    # set_origin: 
-    #   latitude: 0.0
-    #   longitude: 0.0
-    #   altitude: 0.0
-    # earth_to_map_height: 0.0 # set manually height of the map frame above the earth frame
-
