+++ as2_state_estimator/tests/CMakeLists.txt
@@ -7,7 +7,7 @@ foreach(TEST_FILE ${TEST_SOURCE})
 
     add_executable(${PROJECT_NAME}_${TEST_NAME} ${TEST_FILE})
     ament_target_dependencies(${PROJECT_NAME}_${TEST_NAME} ${PROJECT_DEPENDENCIES})
-    target_link_libraries(${PROJECT_NAME}_${TEST_NAME} ${PROJECT_NAME} ${PLUGIN_NAME})
+    target_link_libraries(${PROJECT_NAME}_${TEST_NAME} ${PROJECT_NAME})
 endforeach()
 endif()
 
@@ -22,7 +22,7 @@ foreach(TEST_SOURCE ${GTEST_SOURCE})
 
     ament_add_gtest(${PROJECT_NAME}_${TEST_NAME} ${TEST_SOURCE})
     ament_target_dependencies(${PROJECT_NAME}_${TEST_NAME} ${PROJECT_DEPENDENCIES})
-    target_link_libraries(${PROJECT_NAME}_${TEST_NAME} gtest_main ${PROJECT_NAME} ${PLUGIN_NAME})
+    target_link_libraries(${PROJECT_NAME}_${TEST_NAME} gtest_main ${PROJECT_NAME})
 endforeach()
 endif()
 
@@ -36,6 +36,6 @@ foreach(BENCHMARK_FILE ${BENCHMARK_SOURCE})
     get_filename_component(BENCHMARK_NAME ${BENCHMARK_FILE} NAME_WE)
 
     add_executable(${PROJECT_NAME}_${BENCHMARK_NAME} ${BENCHMARK_FILE})
-    target_link_libraries(${PROJECT_NAME}_${BENCHMARK_NAME} ${PROJECT_NAME} ${PLUGIN_NAME} benchmark::benchmark)
+    target_link_libraries(${PROJECT_NAME}_${BENCHMARK_NAME} ${PROJECT_NAME} benchmark::benchmark)
 endforeach()
 endif()
